2022.10.10

+++ src/main/java/com/example/chattingback/eneity/chat/GroupDto.java
@@ -19,5 +19,5 @@ public class GroupDto {
 
     private GroupMessageDto GroupMessageDto[];
 
-    private int createTime;
+    private long createTime;
 }
